backend/requirements.txt
#############################
# FASTAPI + Server
#############################
fastapi==0.111.0              # Web API framework
uvicorn[standard]==0.29.0     # ASGI server for FastAPI

#############################
# DATABASE (Async PostgreSQL)
#############################
SQLAlchemy==2.0.29            # ORM (async supported)
asyncpg==0.29.0               # Async PostgreSQL driver for SQLAlchemy
# psycopg2-binary removed — NOT needed (sync driver)

#############################
# REDIS + QUEUE
#############################
redis==5.0.4                  # Redis client (sync/async)
rq==1.15.1                    # Job Queue worker

#############################
# ENV + UTILITIES
#############################
python-dotenv==1.0.1          # Load .env environment variables
requests==2.31.0              # HTTP client (rarely used but OK)
pydantic==2.7.0               # Data validation for FastAPI models
aiofiles==23.2.1              # Async file handling (optional but fine)

#############################
# AWS SDK
#############################
boto3==1.34.40                # AWS SDK (S3, DynamoDB, EC2)
botocore==1.34.40             # Core library required by boto3
